Machine Learning Model (RNN) for network traffic
The machine learning model analyzes the different types of network traffic and classifies them accordingly depending on the parameters that are set on the model functions. It analyzes the .csv file with the network traffic data tabulated in the columns necessary for the analysis, once the model is exported, it would be deployed on the web-application for testing and verification. (Subject for improvements)